+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+#### Added
+
+- Reducer: `quantities` for counting unique occurrences of streamed items.
+
 #### Fixed
 
 - A forgotten `all?` export.